fixed automake so builds without Player

+++ libstage/gl.cc
@@ -28,13 +28,13 @@ void Stg::gl_draw_grid( stg_bounds3d_t vol )
 {
   glBegin(GL_LINES);
   
-  for( int i = floor(vol.x.min); i < vol.x.max; i++)
+  for( double i = floor(vol.x.min); i < vol.x.max; i++)
     {
       glVertex2f(  i, vol.y.min );
       glVertex2f(  i, vol.y.max );
     }
   
-  for( int i = floor(vol.y.min); i < vol.y.max; i++)
+  for( double i = floor(vol.y.min); i < vol.y.max; i++)
     {
       glVertex2f(  vol.x.min, i );
       glVertex2f(  vol.x.max, i );
@@ -44,13 +44,13 @@ void Stg::gl_draw_grid( stg_bounds3d_t vol )
   
   char str[16];
 
-  for( int i = floor(vol.x.min); i < vol.x.max; i++)
+  for( double i = floor(vol.x.min); i < vol.x.max; i++)
     {
       snprintf( str, 16, "%d", (int)i );
       gl_draw_string(  i, 0, 0.01, str );
     }
 
-  for( int i = floor(vol.y.min); i < vol.y.max; i++)
+  for( double i = floor(vol.y.min); i < vol.y.max; i++)
     {
       snprintf( str, 16, "%d", (int)i );
       gl_draw_string(  0, i, 0.01, str );
